factor out the creational stuff into a builder, add tests
