+++ profiler/src/test/java/com/splunk/opentelemetry/profiler/ProfilingSupervisorTest.java
@@ -19,20 +19,18 @@ package com.splunk.opentelemetry.profiler;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.awaitility.Awaitility.await;
 import static org.mockito.ArgumentMatchers.anyInt;
+import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.never;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
 import io.opentelemetry.sdk.autoconfigure.AutoConfiguredOpenTelemetrySdk;
-import java.lang.reflect.Field;
-import java.nio.file.Files;
+import io.opentelemetry.sdk.resources.Resource;
 import java.nio.file.Path;
 import java.time.Duration;
-import java.util.Map;
 import java.util.concurrent.ExecutorService;
 import java.util.concurrent.Executors;
-import java.util.concurrent.ScheduledExecutorService;
-import java.util.concurrent.atomic.AtomicReference;
+import java.util.concurrent.LinkedBlockingQueue;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
@@ -45,22 +43,25 @@ import org.mockito.junit.jupiter.MockitoExtension;
 class ProfilingSupervisorTest {
 
   @Mock JFR jfr;
-  TestProfilingConfig config;
-  AutoConfiguredOpenTelemetrySdk sdk;
 
+  TestProfilingConfig config;
+  TestRecordingSequencerBuilder builder;
+  AutoConfiguredOpenTelemetrySdk sdk;
   ExecutorService executor;
   ProfilingSupervisor supervisor;
 
   @BeforeEach
-  void setUp() {
-    executor = Executors.newSingleThreadExecutor();
+  void setUp(@TempDir Path tempDir) {
     config = new TestProfilingConfig();
+    config.profilerDirectory = tempDir.toString();
+    builder = new TestRecordingSequencerBuilder();
+    executor = Executors.newSingleThreadExecutor();
     sdk =
         AutoConfiguredOpenTelemetrySdk.builder()
             .disableShutdownHook()
             .addPropertiesSupplier(
                 () ->
-                    Map.of(
+                    java.util.Map.of(
                         "otel.traces.exporter",
                         "none",
                         "otel.metrics.exporter",
@@ -70,11 +71,12 @@ class ProfilingSupervisorTest {
                         "otel.service.name",
                         "profiling-supervisor-test"))
             .build();
+    supervisor = new TestProfilingSupervisor(config, jfr, sdk, builder);
+    supervisor.start(executor);
   }
 
   @AfterEach
   void tearDown() {
-    shutdownRecordingSequencer();
     executor.shutdownNow();
   }
 
@@ -82,144 +84,81 @@ class ProfilingSupervisorTest {
   void requestStartDoesNotStartProfilerWhenJfrIsUnavailable() {
     when(jfr.isAvailable()).thenReturn(false);
 
-    supervisor = createSupervisor();
-
     supervisor.requestStart();
 
     await().untilAsserted(() -> verify(jfr).isAvailable());
     assertThat(config.logCalled).isFalse();
     verify(jfr, never()).setStackDepth(anyInt());
-    assertThat(currentSequencer()).isNull();
+    assertThat(builder.buildCalled).isFalse();
   }
 
   @Test
-  void requestStartBuildsAndStartsRecordingSequencer(@TempDir Path tempDir) {
-    config.profilerDirectory = tempDir.toString();
+  void requestStartBuildsAndStartsRecordingSequencer() {
     config.stackDepth = 4321;
-    config.callStackInterval = Duration.ofMillis(123);
     config.recordingDuration = Duration.ofMinutes(1);
     when(jfr.isAvailable()).thenReturn(true);
 
-    supervisor = createSupervisor();
-
     supervisor.requestStart();
 
-    RecordingSequencer sequencer = awaitSequencer();
-    JfrRecorder recorder = fieldValue(sequencer, "recorder");
-    Duration actualRecordingDuration = fieldValue(sequencer, "recordingDuration");
-    Duration actualMaxAgeDuration = fieldValue(recorder, "maxAgeDuration");
-    JFR actualJfr = fieldValue(recorder, "jfr");
-    boolean actualKeepRecordingFiles = fieldValue(recorder, "keepRecordingFiles");
-    JfrRecorder actualRecorder = fieldValue(sequencer, "recorder");
-    Object recording = fieldValue(recorder, "recording");
-
+    await().untilAsserted(() -> assertThat(builder.buildCalled).isTrue());
     assertThat(config.logCalled).isTrue();
-    verify(jfr).setStackDepth(4321);
-    assertThat(actualRecordingDuration).isEqualTo(Duration.ofMinutes(1));
-    assertThat(actualJfr).isSameAs(jfr);
-    assertThat(actualMaxAgeDuration).isEqualTo(Duration.ofMinutes(1).multipliedBy(10));
-    assertThat(actualKeepRecordingFiles).isFalse();
-    assertThat(actualRecorder).isSameAs(recorder);
-    assertThat(recording).isNotNull();
-
-    Map<String, String> settings = fieldValue(recorder, "settings");
-    assertThat(settings).containsEntry("jdk.ThreadDump#period", "123 ms");
+    assertThat(builder.jfr).isSameAs(jfr);
+    assertThat(builder.config).isSameAs(config);
+    assertThat(builder.resource).isNotNull();
+    verify(builder.sequencer).start();
   }
 
   @Test
-  void requestStartOnlyStartsOnce(@TempDir Path tempDir) {
-    config.profilerDirectory = tempDir.toString();
+  void requestStartOnlyStartsOnce() {
     when(jfr.isAvailable()).thenReturn(true);
 
-    supervisor = createSupervisor();
-
     supervisor.requestStart();
-    RecordingSequencer sequencer = awaitSequencer();
+    await().untilAsserted(() -> assertThat(builder.buildCalled).isTrue());
     supervisor.requestStart();
 
-    await()
-        .during(Duration.ofMillis(200))
-        .untilAsserted(() -> assertThat(currentSequencer()).isSameAs(sequencer));
-    assertThat(config.logCalled).isTrue();
-    verify(jfr).setStackDepth(1024);
+    await().during(Duration.ofMillis(200)).untilAsserted(() -> verify(builder.sequencer).start());
+    assertThat(builder.buildCount).isEqualTo(1);
   }
 
-  @Test
-  void requestStartCreatesConfiguredOutputDirectoryWhenKeepingFiles(@TempDir Path tempDir) {
-    Path outputDir = tempDir.resolve("profiling-output");
-    config.profilerDirectory = outputDir.toString();
-    config.callStackInterval = Duration.ZERO;
-    config.keepFiles = true;
-    when(jfr.isAvailable()).thenReturn(true);
+  private static class TestProfilingSupervisor extends ProfilingSupervisor {
+    private final RecordingSequencerBuilder builder;
 
-    supervisor = createSupervisor();
-
-    supervisor.requestStart();
-
-    JfrRecorder recorder = fieldValue(awaitSequencer(), "recorder");
-    boolean actualKeepRecordingFiles = fieldValue(recorder, "keepRecordingFiles");
-    assertThat(outputDir).isDirectory();
-    assertThat(actualKeepRecordingFiles).isTrue();
-  }
-
-  @Test
-  void requestStartDisablesKeepingFilesWhenOutputPathIsNotDirectory(@TempDir Path tempDir)
-      throws Exception {
-    Path outputPath = tempDir.resolve("not-a-directory");
-    Files.writeString(outputPath, "already a file");
-    config.profilerDirectory = outputPath.toString();
-    config.callStackInterval = Duration.ZERO;
-    config.keepFiles = true;
-    when(jfr.isAvailable()).thenReturn(true);
-
-    supervisor = createSupervisor();
-
-    supervisor.requestStart();
-
-    JfrRecorder recorder = fieldValue(awaitSequencer(), "recorder");
-    boolean actualKeepRecordingFiles = fieldValue(recorder, "keepRecordingFiles");
-    assertThat(actualKeepRecordingFiles).isFalse();
-  }
-
-  private ProfilingSupervisor createSupervisor() {
-    return ProfilingSupervisor.createAndStart(config, jfr, executor, sdk);
-  }
-
-  private RecordingSequencer awaitSequencer() {
-    await().untilAsserted(() -> assertThat(currentSequencer()).isNotNull());
-    return currentSequencer();
-  }
-
-  private RecordingSequencer currentSequencer() {
-    AtomicReference<RecordingSequencer> sequencerReference = fieldValue(supervisor, "sequencer");
-    return sequencerReference.get();
-  }
-
-  private void shutdownRecordingSequencer() {
-    if (supervisor == null) {
-      return;
+    TestProfilingSupervisor(
+        ProfilerConfiguration config,
+        JFR jfr,
+        AutoConfiguredOpenTelemetrySdk sdk,
+        RecordingSequencerBuilder builder) {
+      super(config, jfr, sdk, new LinkedBlockingQueue<>());
+      this.builder = builder;
     }
-    AtomicReference<RecordingSequencer> sequencerReference = fieldValue(supervisor, "sequencer");
-    RecordingSequencer sequencer = sequencerReference.get();
-    if (sequencer != null) {
-      ScheduledExecutorService sequencerExecutor = fieldValue(sequencer, "executor");
-      sequencerExecutor.shutdownNow();
-      JfrRecorder recorder = fieldValue(sequencer, "recorder");
-      Object recording = fieldValue(recorder, "recording");
-      if (recording != null) {
-        recorder.stop();
-      }
+
+    @Override
+    RecordingSequencerBuilder makeRecordingSequencerBuilder() {
+      return builder;
     }
   }
 
-  @SuppressWarnings("unchecked")
-  private static <T> T fieldValue(Object target, String fieldName) {
-    try {
-      Field field = target.getClass().getDeclaredField(fieldName);
-      field.setAccessible(true);
-      return (T) field.get(target);
-    } catch (ReflectiveOperationException e) {
-      throw new AssertionError("Could not read field " + fieldName + " from " + target, e);
+  private static class TestRecordingSequencerBuilder extends RecordingSequencerBuilder {
+    final RecordingSequencer sequencer = mock(RecordingSequencer.class);
+    JFR jfr;
+    ProfilerConfiguration config;
+    Resource resource;
+    boolean buildCalled;
+    int buildCount;
+
+    @Override
+    RecordingSequencerBuilder jfr(JFR jfr) {
+      this.jfr = jfr;
+      return this;
+    }
+
+    @Override
+    RecordingSequencer build(ProfilerConfiguration config, Resource resource) {
+      this.config = config;
+      this.resource = resource;
+      buildCalled = true;
+      buildCount++;
+      return sequencer;
     }
   }
 }
